API CHANGE: Use MKBuild and update to mkmmdb@v0.4.0
This diff uses MKBuild to generate build files. As part of that
we have upgraded to mkmmdb@v0.4.0, whose API changed. More in
detail, v0.4.0 of mkmmdb now has a different method name to get
the ASN because now the ASN is a string prefixed with `"AS"`.

The method name changed to signal this different. Likewise, also
here in mkgeoip let's change the name of the structure field in
which the ASN string is copied. This will break the build and so
more clearly signal that something has really changed.